data-engineering/financial-reporting-platform/requirements.txt
pandas==2.0.3
numpy==1.24.3
pyspark==3.5.0
delta-spark==2.4.0
duckdb==0.9.2
plotly==5.17.0
dash==2.14.1
streamlit==1.28.0
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
faker==20.0.0
pyyaml==6.0
pytest==7.4.3
prefect==2.14.5
fastapi==0.104.1
uvicorn[standard]==0.24.0
openpyxl==3.1.2